too many changes, I create pdf!

+++ diving/lib/Models/Course.dart
@@ -9,11 +9,11 @@ class Course{
   Course(this.id,this.courseName,this.cost,this.description,this.minHoursUnderWater,this.image);
 
   Course.fromJson(Map<String, dynamic> json){
-    id = json["id"];
+    id = json["id"] as int;
     courseName = json["courseName"];
-    cost = json["cost"];
+    cost = json["cost"] as double;
     description = json["description"];
-    minHoursUnderWater = json["minHoursUnderWater"];
+    minHoursUnderWater = json["minHoursUnderWater"] as int;
     image = json["image"];
   }
